Refactor client into HTTP and WebSocket components
